Remove the notion of view changing ops
Summary:
These ops are now treated as inplace and implemented as such.

LSTM training now passes - the error seems like an accumulation of fp32 error as the absolute error is small (even with a big learning rate).

Fixes T5469

Test Plan: All tests pass - note that some tests required changing the names of compute sets but no actual new copies have been added (for example adds have been turned into a copy with an AddTo) - including training and speed/size suites. Also checked the CE examples I have been running last week.

Reviewers: jakeh, davidn

Reviewed By: davidn

Maniphest Tasks: T5469

Differential Revision: https://phabricator.sourcevertex.net/D7592

+++ tensorflow/compiler/plugin/poplar/driver/visitor_inline_call.cc
@@ -28,7 +28,7 @@ InlineCallVisitor::InlineCallVisitor(CompilerResources& res,
     : FullVisitor(res), inputs_(std::move(inputs)) {}
 
 Status InlineCallVisitor::HandleParameter(HloInstruction* inst) {
-  for (unsigned int t = 0; t < inputs_[inst->parameter_number()].size(); t++) {
+  for (uint64 t = 0; t < inputs_[inst->parameter_number()].size(); t++) {
     auto& v = inputs_[inst->parameter_number()];
     TF_CHECK_OK(AddOutputTensor(tensor_map, inst, t, v[t]));
   }
